+++ demos/src/Examples/Community/Vue/index.vue
@@ -1,41 +1,39 @@
 <template>
-  <div>
-    <editor-content :editor="editor" />
+  <editor-content :editor="editor" />
 
-    <div v-if="editor" :class="{'character-count': true, 'character-count--warning': editor.storage.characterCount.characters() === limit}">
-      <svg
-        height="20"
-        width="20"
-        viewBox="0 0 20 20"
-        class="character-count__graph"
-      >
-        <circle
-          r="10"
-          cx="10"
-          cy="10"
-          fill="#e9ecef"
-        />
-        <circle
-          r="5"
-          cx="10"
-          cy="10"
-          fill="transparent"
-          stroke="currentColor"
-          stroke-width="10"
-          :stroke-dasharray="`calc(${percentage} * 31.4 / 100) 31.4`"
-          transform="rotate(-90) translate(-20)"
-        />
-        <circle
-          r="6"
-          cx="10"
-          cy="10"
-          fill="white"
-        />
-      </svg>
+  <div v-if="editor" :class="{'character-count': true, 'character-count--warning': editor.storage.characterCount.characters() === limit}">
+    <svg
+      height="20"
+      width="20"
+      viewBox="0 0 20 20"
+      class="character-count__graph"
+    >
+      <circle
+        r="10"
+        cx="10"
+        cy="10"
+        fill="#e9ecef"
+      />
+      <circle
+        r="5"
+        cx="10"
+        cy="10"
+        fill="transparent"
+        stroke="currentColor"
+        stroke-width="10"
+        :stroke-dasharray="`calc(${percentage} * 31.4 / 100) 31.4`"
+        transform="rotate(-90) translate(-20)"
+      />
+      <circle
+        r="6"
+        cx="10"
+        cy="10"
+        fill="white"
+      />
+    </svg>
 
-      <div class="character-count__text">
-        {{ editor.storage.characterCount.characters() }}/{{ limit }} characters
-      </div>
+    <div class="character-count__text">
+      {{ editor.storage.characterCount.characters() }}/{{ limit }} characters
     </div>
   </div>
 </template>