+++ packages/core/src/commands/wrapIn.ts
@@ -1,7 +1,6 @@
 import { wrapIn as originalWrapIn } from 'prosemirror-commands'
 import { NodeType } from 'prosemirror-model'
 import { RawCommands } from '../types'
-import { isNodeActive } from '../helpers/isNodeActive'
 import { getNodeType } from '../helpers/getNodeType'
 
 declare module '@tiptap/core' {
@@ -17,11 +16,6 @@ declare module '@tiptap/core' {
 
 export const wrapIn: RawCommands['wrapIn'] = (typeOrName, attributes = {}) => ({ state, dispatch }) => {
   const type = getNodeType(typeOrName, state.schema)
-  const isActive = isNodeActive(state, type, attributes)
-
-  if (isActive) {
-    return false
-  }
 
   return originalWrapIn(type, attributes)(state, dispatch)
 }