+++ packages/extension-character-count/src/character-count.ts
@@ -6,7 +6,7 @@ export interface CharacterCountOptions {
   /**
    * The maximum number of characters that should be allowed. Defaults to `0`.
    */
-  limit: number,
+  limit: number | null | undefined,
   /**
    * The mode by which the size is calculated. Defaults to 'textSize'.
    */
@@ -17,7 +17,7 @@ export interface CharacterCountStorage {
   /**
    * Get the number of characters for the current document.
    */
-  characters?: (options: {
+  characters: (options?: {
     node?: ProseMirrorNode,
     mode?: 'textSize' | 'nodeSize',
   }) => number,
@@ -25,7 +25,7 @@ export interface CharacterCountStorage {
   /**
    * Get the number of words for the current document.
    */
-  words?: (options: {
+  words: (options?: {
     node?: ProseMirrorNode,
   }) => number,
 }
@@ -35,15 +35,15 @@ export const CharacterCount = Extension.create<CharacterCountOptions, CharacterC
 
   addOptions() {
     return {
-      limit: 0,
+      limit: null,
       mode: 'textSize',
     }
   },
 
   addStorage() {
     return {
-      characters: undefined,
-      words: undefined,
+      characters: () => 0,
+      words: () => 0,
     }
   },
 
@@ -80,12 +80,12 @@ export const CharacterCount = Extension.create<CharacterCountOptions, CharacterC
           const limit = this.options.limit
 
           // Nothing has changed or no limit is defined. Ignore it.
-          if (!transaction.docChanged || limit === 0) {
+          if (!transaction.docChanged || limit === 0 || limit === null || limit === undefined) {
             return true
           }
 
-          const oldSize = this.storage.characters?.({ node: state.doc }) || 0
-          const newSize = this.storage.characters?.({ node: transaction.doc }) || 0
+          const oldSize = this.storage.characters({ node: state.doc })
+          const newSize = this.storage.characters({ node: transaction.doc })
 
           // Everything is in the limit. Good.
           if (newSize <= limit) {
@@ -123,7 +123,7 @@ export const CharacterCount = Extension.create<CharacterCountOptions, CharacterC
           // This happens e.g. when truncating within a complex node (e.g. table)
           // and ProseMirror has to close this node again.
           // If this is the case, we prevent the transaction completely.
-          const updatedSize = this.storage.characters?.({ node: transaction.doc }) || 0
+          const updatedSize = this.storage.characters({ node: transaction.doc })
 
           if (updatedSize > limit) {
             return false